testing oferta esta genial

+++ database/migrations/2025_04_26_122205_noobligatorios.php
@@ -4,23 +4,24 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
-    {
-        Schema::table('equipos', function (Blueprint $table) {
-			$table->dropColumn('Clasificacion 2 Inventario');
+return new class extends Migration {
+	
+	/**
+	 * Run the migrations.
+	 */
+	public function up() {
+		Schema::table('equipos', function (Blueprint $table) {
+			if (Schema::hasColumn('equipos', 'Clasificacion 2 Inventario')) {
+				$table->dropColumn('Clasificacion 2 Inventario');
+			}
 		});
-    }
-
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
-    {
-        //
-    }
+	}
+	
+	public function down() {
+		Schema::table('equipos', function (Blueprint $table) {
+			// Si necesitas revertir, puedes agregar la columna de nuevo aquí.
+			// Por ejemplo:
+			// $table->string('Clasificacion 2 Inventario')->nullable();
+		});
+	}
 };
